agregado guards para proteger rutas

+++ src/app/pages/amigos/amigos.component.ts
@@ -93,10 +93,7 @@ export class AmigosComponent {
 
   unfollowUser(idToUnfollow: string){
     console.log(idToUnfollow)
-    let elementPos = this.followed.map(function(x) {return x.id; }).indexOf(idToUnfollow);
-    let objectFound = this.followed[elementPos]
-    console.log(elementPos)
-    /* if(this.user){
+    if(this.user){
       this.amigosService.deleteFollow(this.user.id, idToUnfollow).subscribe(
         (response)=>{
           this.updatefollowed()
@@ -106,7 +103,7 @@ export class AmigosComponent {
           console.log(error)
         }
       )
-    } */
+    }
   }
 
   updatefollowed(){
